Backend/brequirements.txt
asgiref
Django
django-cors-headers # For handling Cross-Origin Resource Sharing (CORS) in Django applications  
djangorestframework
djangorestframework-simplejwt # JWT authentication for Django REST Framework
PyJWT
pytz
sqlparse
psycopg2-binary # PostgreSQL database adapter for Python
python-dotenv # For loading environment variables from a .env file
django-allauth # Social auth (Google, etc.)
requests # HTTP client used by allauth Google provider
cryptography # Required by pyjwt for secure token handling